requirements.txt
numpy
opencv-python
Pillow
pytest
planetary-computer
pystac-client
rasterio
geopandas
shapely
tqdm